rename user to userReducer
