+++ packages/client/src/app/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { userReducer } from './reducers/user';
+import { userReducer } from './reducers/userReducer';
 
 export const store = configureStore({
     reducer: { user: userReducer },
